Add basic server audit tab to system console for EE
