+++ web/static/i18n/en.json
@@ -8,53 +8,54 @@
   "about.date": "Build Date:",
   "about.hash": "Build Hash:",
   "about.close": "Close",
-  "access_history.sessionRevoked": "The session with id {sessionId} was revoked",
-  "access_history.channelCreated": "Created the {channelName} channel/group",
-  "access_history.establishedDM": "Established a direct message channel with {username}",
-  "access_history.nameUpdated": "Updated the {channelName} channel/group name",
-  "access_history.headerUpdated": "Updated the {channelName} channel/group header",
-  "access_history.channelDeleted": "Deleted the channel/group with the URL {url}",
-  "access_history.userAdded": "Added {username} to the {channelName} channel/group",
-  "access_history.userRemoved": "Removed {username} to the {channelName} channel/group",
-  "access_history.attemptedRegisterApp": "Attempted to register a new OAuth Application with ID {id}",
-  "access_history.attemptedAllowOAuthAccess": "Attempted to allow a new OAuth service access",
-  "access_history.successfullOAuthAccess": "Successfully gave a new OAuth service access",
-  "access_history.failedOAuthAccess": "Failed to allow a new OAuth service access - the redirect URI did not match the previously registered callback",
-  "access_history.attemptedOAuthToken": "Attempted to get an OAuth access token",
-  "access_history.successfullOAuthToken": "Successfully added a new OAuth service",
-  "access_history.oauthTokenFailed": "Failed to get an OAuth access token - {token}",
-  "access_history.attemptedLogin": "Attempted to login",
-  "access_history.successfullLogin": "Successfully logged in",
-  "access_history.failedLogin": "FAILED login attempt",
-  "access_history.updatePicture": "Updated your profile picture",
-  "access_history.updateGeneral": "Updated the general settings of your account",
-  "access_history.attemptedPassword": "Attempted to change password",
-  "access_history.successfullPassword": "Successfully changed password",
-  "access_history.failedPassword": "Failed to change password - tried to update user password who was logged in through oauth",
-  "access_history.updatedRol": "Updated user role(s) to ",
-  "access_history.member": "member",
-  "access_history.accountActive": "Account made active",
-  "access_history.accountInactive": "Account made inactive",
-  "access_history.by": " by {username}",
-  "access_history.byAdmin": " by an admin",
-  "access_history.sentEmail": "Sent an email to {email} to reset your password",
-  "access_history.attemptedReset": "Attempted to reset password",
-  "access_history.successfullReset": "Successfully reset password",
-  "access_history.updateGlobalNotifications": "Updated your global notification settings",
-  "access_history.attemptedWebhookCreate": "Attempted to create a webhook",
-  "access_history.successfullWebhookCreate": "Successfully created a webhook",
-  "access_history.failedWebhookCreate": "Failed to create a webhook - bad channel permissions",
-  "access_history.attemptedWebhookDelete": "Attempted to delete a webhook",
-  "access_history.successfullWebhookDelete": "Successfully deleted a webhook",
-  "access_history.failedWebhookDelete": "Failed to delete a webhook - inappropriate conditions",
-  "access_history.logout": "Logged out of your account",
-  "access_history.verified": "Sucessfully verified your email address",
-  "access_history.revokedAll": "Revoked all current sessions for the team",
-  "access_history.loginAttempt": " (Login attempt)",
-  "access_history.loginFailure": " (Login failure)",
-  "access_history.moreInfo": "More info",
-  "access_history.ip": "IP: {ip}",
-  "access_history.session": "Session ID: {id}",
+  "audit_table.sessionRevoked": "The session with id {sessionId} was revoked",
+  "audit_table.channelCreated": "Created the {channelName} channel/group",
+  "audit_table.establishedDM": "Established a direct message channel with {username}",
+  "audit_table.nameUpdated": "Updated the {channelName} channel/group name",
+  "audit_table.headerUpdated": "Updated the {channelName} channel/group header",
+  "audit_table.channelDeleted": "Deleted the channel/group with the URL {url}",
+  "audit_table.userAdded": "Added {username} to the {channelName} channel/group",
+  "audit_table.userRemoved": "Removed {username} to the {channelName} channel/group",
+  "audit_table.attemptedRegisterApp": "Attempted to register a new OAuth Application with ID {id}",
+  "audit_table.attemptedAllowOAuthAccess": "Attempted to allow a new OAuth service access",
+  "audit_table.successfullOAuthAccess": "Successfully gave a new OAuth service access",
+  "audit_table.failedOAuthAccess": "Failed to allow a new OAuth service access - the redirect URI did not match the previously registered callback",
+  "audit_table.attemptedOAuthToken": "Attempted to get an OAuth access token",
+  "audit_table.successfullOAuthToken": "Successfully added a new OAuth service",
+  "audit_table.oauthTokenFailed": "Failed to get an OAuth access token - {token}",
+  "audit_table.attemptedLogin": "Attempted to login",
+  "audit_table.successfullLogin": "Successfully logged in",
+  "audit_table.failedLogin": "FAILED login attempt",
+  "audit_table.updatePicture": "Updated your profile picture",
+  "audit_table.updateGeneral": "Updated the general settings of your account",
+  "audit_table.attemptedPassword": "Attempted to change password",
+  "audit_table.successfullPassword": "Successfully changed password",
+  "audit_table.failedPassword": "Failed to change password - tried to update user password who was logged in through oauth",
+  "audit_table.updatedRol": "Updated user role(s) to ",
+  "audit_table.member": "member",
+  "audit_table.accountActive": "Account made active",
+  "audit_table.accountInactive": "Account made inactive",
+  "audit_table.by": " by {username}",
+  "audit_table.byAdmin": " by an admin",
+  "audit_table.sentEmail": "Sent an email to {email} to reset your password",
+  "audit_table.attemptedReset": "Attempted to reset password",
+  "audit_table.successfullReset": "Successfully reset password",
+  "audit_table.updateGlobalNotifications": "Updated your global notification settings",
+  "audit_table.attemptedWebhookCreate": "Attempted to create a webhook",
+  "audit_table.successfullWebhookCreate": "Successfully created a webhook",
+  "audit_table.failedWebhookCreate": "Failed to create a webhook - bad channel permissions",
+  "audit_table.attemptedWebhookDelete": "Attempted to delete a webhook",
+  "audit_table.successfullWebhookDelete": "Successfully deleted a webhook",
+  "audit_table.failedWebhookDelete": "Failed to delete a webhook - inappropriate conditions",
+  "audit_table.logout": "Logged out of your account",
+  "audit_table.verified": "Sucessfully verified your email address",
+  "audit_table.revokedAll": "Revoked all current sessions for the team",
+  "audit_table.loginAttempt": " (Login attempt)",
+  "audit_table.loginFailure": " (Login failure)",
+  "audit_table.moreInfo": "More info",
+  "audit_table.ip": "IP: {ip}",
+  "audit_table.session": "Session ID: {id}",
+  "audit_table.userId": "User ID",
   "access_history.title": "Access History",
   "activity_log_modal.iphoneNativeApp": "iPhone Native App",
   "activity_log_modal.androidNativeApp": "Android Native App",
@@ -96,6 +97,7 @@
   "admin.sidebar.teams": "TEAMS ({count})",
   "admin.sidebar.other": "OTHER",
   "admin.sidebar.logs": "Logs",
+  "admin.sidebar.audits": "Audits",
   "admin.analytics.loading": "Loading...",
   "admin.analytics.totalUsers": "Total Users",
   "admin.analytics.publicChannels": "Public Channels",
@@ -326,6 +328,8 @@
   "admin.log.save": "Save",
   "admin.logs.title": "Server Logs",
   "admin.logs.reload": "Reload",
+  "admin.audits.title": "Server Audits",
+  "admin.audits.reload": "Reload",
   "admin.privacy.saving": "Saving Config...",
   "admin.privacy.title": "Privacy Settings",
   "admin.privacy.showEmailTitle": "Show Email Address: ",
